requirements.txt
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
pep8==1.7.1
Pillow==6.0.0
pycodestyle==2.5.0
pydocstyle==3.0.0
PyJWT==1.7.1
pytz==2019.1
requests==2.22.0
six==1.12.0
snowballstemmer==1.2.1
sqlparse==0.3.0
urllib3==1.25.3
